Fixed authenticated_user_id requirement on relevant tools. Added Tool disable config. TTS fixed number conversion and rupee symbol. Changed system Prompt

+++ openg2p-llm-common/src/openg2p_llm_common/config.py
@@ -9,7 +9,7 @@ from . import __version__
 
 class OllamaOptions(BaseModel):
     model_config = ConfigDict(extra="allow")
-    temperature: int = 0
+    temperature: float = 0
 
 
 class Settings(BaseSettings):
